Created server to serve mysql → Redis syncronization requests
